+++ src/main/java/com/sw/ecogrowbackend/domain/community/controller/PostController.java
@@ -1,8 +1,8 @@
-package com.sw.ecogrowbackend.community.controller;
+package com.sw.ecogrowbackend.domain.community.controller;
 
-import com.sw.ecogrowbackend.community.dto.PostDto;
-import com.sw.ecogrowbackend.community.entity.Post;
-import com.sw.ecogrowbackend.community.service.PostService;
+import com.sw.ecogrowbackend.domain.community.dto.PostDto;
+import com.sw.ecogrowbackend.domain.community.entity.Post;
+import com.sw.ecogrowbackend.domain.community.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
