+++ src/main/java/com/sw/ecogrowbackend/domain/community/dto/PostDto.java
@@ -1,4 +1,4 @@
-package com.sw.ecogrowbackend.community.dto;
+package com.sw.ecogrowbackend.domain.community.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CommentDto {
+public class PostDto {
+    private String title;
     private String content;
+    private boolean isNotice;
 }
