+++ src/main/java/com/sw/ecogrowbackend/domain/community/service/PostService.java
@@ -1,10 +1,10 @@
-package com.sw.ecogrowbackend.community.service;
+package com.sw.ecogrowbackend.domain.community.service;
 
 import com.sw.ecogrowbackend.common.exception.CustomException;
 import com.sw.ecogrowbackend.common.exception.ErrorCode;
-import com.sw.ecogrowbackend.community.dto.PostDto;
-import com.sw.ecogrowbackend.community.entity.Post;
-import com.sw.ecogrowbackend.community.repository.PostRepository;
+import com.sw.ecogrowbackend.domain.community.dto.PostDto;
+import com.sw.ecogrowbackend.domain.community.entity.Post;
+import com.sw.ecogrowbackend.domain.community.repository.PostRepository;
 import com.sw.ecogrowbackend.domain.profile.entity.Profile;
 import com.sw.ecogrowbackend.domain.profile.repository.ProfileRepository;
 import lombok.RequiredArgsConstructor;
